+++ navigation-compose/src/main/kotlin/com/freeletics/khonshu/navigation/androidx/internal/StoreViewModel.kt
@@ -1,6 +1,8 @@
-package com.freeletics.khonshu.navigation.internal
+package com.freeletics.khonshu.navigation.androidx.internal
 
 import androidx.lifecycle.ViewModel
+import com.freeletics.khonshu.navigation.internal.NavigationExecutor
+import com.freeletics.khonshu.navigation.internal.NavigationExecutorStore
 
 internal class StoreViewModel private constructor(
     store: NavigationExecutorStore,